requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
fastapi==0.95.1
idna==3.4
numpy==1.24.3
pandas==2.0.1
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
